Fix import key size
This commit fixes the key size set by in the import operation in the
abstract test client. Previously, the value was set to the length of
the data passed as the key to be imported, in bytes, which is incorrect
in two ways:
* The length should be in bits
* The length should cover only the key, not the surrounding metadata as
well
Setting it to 0 is correct in the current definition of the PSA spec, as
the `data_len` field passed in the operation is enough.

Signed-off-by: Ionut Mihalcea <ionut.mihalcea@arm.com>